Add files via upload
This repository contains the code for a simple to do list application that uses JavaScript, HTML, and CSS. The application allows users to create, edit, and delete tasks. Tasks can be marked as complete, and the application will save the tasks to local storage so that they are not lost when the browser is closed.